+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.4] - 2025-01-XX
+
+### Fixed
+
+- Fixed ZIM files not being added to library.xml after download/upload
+- Management API now uses `kiwix-manage` to add/remove ZIM files from library.xml
+- ZIM files are now automatically detected by Kiwix server after download/upload
+- Fixed library.xml not being updated when files are added
+
 ## [1.1.3] - 2025-01-XX
 
 ### Fixed